Feature/kevin rojas

+++ ABServiceTransaction/src/main/java/com/nttdata/abs/transaction/repository/TypeRepository.java
@@ -4,6 +4,6 @@ import com.nttdata.abs.transaction.entity.Type;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TypeRepository extends JpaRepository<Type, Long> {
+public interface TypeRepository extends JpaRepository<Type, Integer> {
 
 }
